packages/sdl-schema/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "usf-sdl-schema"
version = "0.1.0"
description = "Pydantic v2 models and validator for the USF Semantic Definition Language (SDL)"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.7",
    "pyyaml>=6.0",
    "rdflib>=7.0",
]
[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4", "mypy>=1.10"]
[tool.hatch.build.targets.wheel]
packages = ["usf_sdl"]
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
